requirements.txt
google-generativeai>=0.2.0
python-dotenv>=1.0.0
Pillow>=9.0.0
keyboard>=0.13.5